More css fixes and organization
